soc/google/cpif: apply iface and ch mod for dit 2.2
DIT 2.2 added DIT_REG_NAT_INTERFACE_NUM register and an interface
field at desc. If DIT_REG_NAT_INTERFACE_NUM is zero, every packets
bypass HW filter. Should set DIT_REG_NAT_INTERFACE_NUM when tethering
is enabled.

It does not need to add DIT_REG_NAT_INTERFACE_NUM to backup/restore
items due to memory retention from DIT 2.2. Set the related functions
to DIT 2.1 specific features.

An interface field can replace the ch_id defined by SW. Use the
interface as ch_id.

+++ drivers/soc/google/cpif/dit.h
@@ -40,7 +40,7 @@ extern u32 gs_chipid_get_type(void);
 #if IS_ENABLED(CONFIG_EXYNOS_DIT)
 int dit_enqueue_src_desc_ring(
 	enum dit_direction dir, u8 *src, unsigned long src_paddr,
-	u16 len, u16 ch_id, bool csum);
+	u16 len, u8 ch_id, bool csum);
 int dit_enqueue_src_desc_ring_skb(enum dit_direction dir, struct sk_buff *skb);
 int dit_kick(enum dit_direction dir, bool retry);
 bool dit_check_dir_use_queue(enum dit_direction dir, unsigned int queue_num);
@@ -50,7 +50,7 @@ bool dit_support_clat(void);
 #else
 static inline int dit_enqueue_src_desc_ring(
 	enum dit_direction dir, u8 *src, unsigned long src_paddr,
-	u16 len, u16 ch_id, bool csum) { return -1; }
+	u16 len, u8 ch_id, bool csum) { return -1; }
 static inline int dit_enqueue_src_desc_ring_skb(
 	enum dit_direction dir, struct sk_buff *skb) { return -1; }
 static inline int dit_kick(enum dit_direction dir, bool retry) { return -1; }
